+++ authentification_api/src/main/java/mg/itu/auth/service/ValidationCodeService.java
@@ -9,7 +9,7 @@ public class ValidationCodeService {
     private static final SecureRandom random = new SecureRandom();
 
     public int generateValidationCode() {
-        int min = (short) Math.pow(10, CODE_LENGTH - 1); // 100000
+        int min = (int) Math.pow(10, CODE_LENGTH - 1); // 100000
         int max = (int) Math.pow(10, CODE_LENGTH) - 1; // 999999
         return random.nextInt(max - min + 1) + min; // Génère un nombre entre 100000 et 999999
     }
